Add controller spec for catches.

+++ db/migrate/20160911093349_create_catches.rb
@@ -3,7 +3,7 @@ class CreateCatches < ActiveRecord::Migration[5.0]
     create_table :catches do |t|
       t.belongs_to :user
       t.string :species
-      t.decimal :weight
+      t.integer :weight
       t.integer :length
       t.float :lang
       t.float :lat
